pyproject.toml
[tool.poetry]
name = "kr-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
django = "^4.2"
redis = "^5.0.8"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
